backend/requirements.txt
anyio==3.6.2
async-timeout==4.0.2
bcrypt==4.0.1
black==23.1.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
cryptography==39.0.1
dnspython==2.3.0
ecdsa==0.18.0
fastapi==0.89.1
h11==0.14.0
httptools==0.5.0
idna==3.4
mypy-extensions==0.4.3
numpy==1.24.1
packaging==23.0
pandas==1.5.3
passlib==1.7.4
pathspec==0.11.0
platformdirs==2.6.2
pyasn1==0.4.8
pycparser==2.21
pydantic==1.10.4
pymongo==4.3.3
python-dateutil==2.8.2
python-dotenv==0.21.1
python-jose==3.3.0
python-multipart==0.0.5
pytz==2022.7.1
PyYAML==6.0
redis==4.5.1
requests==2.28.2
rsa==4.9
six==1.16.0
sniffio==1.3.0
spotipy==2.22.1
starlette==0.22.0
tomli==2.0.1
typing_extensions==4.4.0
urllib3==1.26.14
uuid==1.30
uvicorn==0.20.0
uvloop==0.17.0
watchfiles==0.18.1
websockets==10.4
